order for consignment, update batch, delete

+++ SWP391-KoiFarmShop/src/main/java/com/koishop/models/user_model/Point.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class Point {
-    private double point;
+    private int point;
 }
